requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
pdfkit==1.0.0
psycopg2-binary==2.9.6
aws-psycopg2==1.3.8
boto3==1.26.105
sendgrid==6.4.8
python-dateutil==2.8.2